refactor!: Rewrite capability support to reduce the need for duplicate code
